+++ gusto_embedded/src/funcs/payrollsGetReceipt.ts
@@ -21,6 +21,10 @@ import {
 import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
 import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
 import {
+  UnprocessableEntityErrorObject,
+  UnprocessableEntityErrorObject$inboundSchema,
+} from "../models/errors/unprocessableentityerrorobject.js";
+import {
   GetV1PaymentReceiptsPayrollsPayrollUuidRequest,
   GetV1PaymentReceiptsPayrollsPayrollUuidRequest$outboundSchema,
   GetV1PaymentReceiptsPayrollsPayrollUuidResponse,
@@ -49,6 +53,7 @@ export function payrollsGetReceipt(
 ): APIPromise<
   Result<
     GetV1PaymentReceiptsPayrollsPayrollUuidResponse,
+    | UnprocessableEntityErrorObject
     | GustoEmbeddedError
     | ResponseValidationError
     | ConnectionError
@@ -74,6 +79,7 @@ async function $do(
   [
     Result<
       GetV1PaymentReceiptsPayrollsPayrollUuidResponse,
+      | UnprocessableEntityErrorObject
       | GustoEmbeddedError
       | ResponseValidationError
       | ConnectionError
@@ -171,6 +177,7 @@ async function $do(
 
   const [result] = await M.match<
     GetV1PaymentReceiptsPayrollsPayrollUuidResponse,
+    | UnprocessableEntityErrorObject
     | GustoEmbeddedError
     | ResponseValidationError
     | ConnectionError
@@ -183,7 +190,8 @@ async function $do(
     M.json(200, GetV1PaymentReceiptsPayrollsPayrollUuidResponse$inboundSchema, {
       key: "Payroll-Receipt",
     }),
-    M.fail([404, "4XX"]),
+    M.jsonErr(404, UnprocessableEntityErrorObject$inboundSchema),
+    M.fail("4XX"),
     M.fail("5XX"),
   )(response, req, { extraFields: responseFields });
   if (!result.ok) {